Separate the section enum into different files
