+++ StoreApp/StoreApp/ViewController.swift
@@ -82,22 +82,3 @@ extension ViewController: UITableViewDelegate {
         return 80
     }
 }
-
-extension ViewController {
-    enum Section: Int {
-        case main, soup, side
-        var title: String {
-            switch self {
-            case .main: return .mainTitle
-            case .soup: return .soupTitle
-            case .side: return .sideTitle }
-        }
-        var detail: String {
-            switch self {
-            case .main: return .mainDetail
-            case .soup: return .soupDetail
-            case .side: return .sideDetail }
-        }
-        static var numberOfSections: Int { return 3 }
-    }
-}
